requirements.txt
beautifulsoup4==4.4.1
decorator==4.0.9
dict==0.0.5
Django==1.9.6
django-markdown-deux==1.0.5
django-markdown2==0.3.0
django-markdownx==1.4.3
get==0.0.7
gunicorn==19.4.5
Markdown==2.6.6
markdown2==2.3.1
Pillow==3.2.0
post==0.0.5
PyMySQL==0.7.2
query-string==0.0.5
requests==2.10.0
self==0.0.5
